[feature] hw_L65: created 'cart.routes.js', replaced cart routes from 'server.js' to 'cart.routes.js', changes applied
